+++ tools/analyzer/pass.h
@@ -31,7 +31,7 @@ protected:
   int64_t visitingDepth{0};
 
 public:
-  virtual bool enterModule(unsigned short modID, int version,
+  virtual bool enterModule(unsigned short modId, int version,
                            const std::shared_ptr<TypePool> &pool) = 0;
   virtual bool exitModule(const std::unique_ptr<Node> &node) = 0;
 
@@ -126,7 +126,7 @@ private:
   std::vector<ObserverPtr<NodePass>> passes;
 
 public:
-  bool enterModule(unsigned short modID, int version,
+  bool enterModule(unsigned short modId, int version,
                    const std::shared_ptr<TypePool> &pool) override;
   bool exitModule(const std::unique_ptr<Node> &node) override;
   bool consume(const std::unique_ptr<Node> &node) override;