+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - ``textDocument/publishDiagnostics`` emit warning diagnostic
     - meaningless cast op
     - variable shadowing
+    - unused local variables
 
 ### Changed